feat: allow outout column selection
Major refactor to define the columns available to print and how to render them. This DRYs-up some duplication between rendering tables and CSV.
Default columns are largely the same as before but now with the "-c" flag you can pick what you want and the order.
This also controls whether or not the MD5 and SHA1 are calcualted in "ls" mode, so the `--with-md5` flag is gone now.